Handle 100% balance orders when price manual
Gitlab issue:
https://gitlab.com/crankykernel/maker/issues/34

When price is set to manual, base the amount of the asset to
buy based on the input price, not the last price.

Also change the amount of quote asset available to be only 99.99%
of whats really available to cover any fees.  Binance will error
out if their is not enough balance left to cover fees.